Fix Door opening once curfew has been lifted
If the phone(s) is away before the curfew is imposed and the phone returns before the curfew is lifted, then once the curfew is lifted, the script will see that it's ok to open the garage door & the phone has arrived. This was fixed by setting the ok_to_open variable to False in the check_curfew function when the curfew is True.