Fix door not opening bug & code cleanup/comments
If the owner leaves and comes back after curfew starts and then leaves again before curfew ends, the script will not open the door when the owner returns home while curfew is off. This appears to be because the check_phones function does not create a new timer if j_away is equal to j_away_previous.